+++ examples/http-certification/custom-assets/src/backend/src/lib.rs
@@ -16,7 +16,7 @@ use std::{cell::RefCell, collections::HashMap};
 
 #[derive(Debug, Clone, Serialize)]
 pub struct Metrics {
-    pub cycle_balance: u64,
+    pub cycle_balance: u128,
 }
 
 // Public methods
@@ -358,7 +358,7 @@ fn asset_handler(req: &HttpRequest) -> HttpResponse<'static> {
 
 fn create_metrics_response() -> HttpResponse<'static> {
     let metrics = Metrics {
-        cycle_balance: canister_cycle_balance() as u64,
+        cycle_balance: canister_cycle_balance(),
     };
     let body = serde_json::to_vec(&metrics).expect("Failed to serialize metrics");
     let additional_headers = vec![
